Added mashup visualizers for forecasting and refactored kinematic state visualizer for mashup support
